pyproject.toml
[project]
name = "makeupai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.25.0",
    "fpdf2>=2.8.5",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
]
